Added Info and Debug Buttons
Added info panels with updating text and colours, Added the buttons to change this info, as well as 4 debug options to test out if the info updates or not.

Should also say buttons now check to see if they have the needed items to do their functions. If so, they do it. If not, they output an error to the console. If the player runs out of bombs/arrows, the info box colour changes from green to red, and stays red until the player gets more of that item, at which case it turns green again.

Also yet another attempt at fixing the button issue. Still ongoing.